% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{counts.per.seg}
\alias{counts.per.seg}
\title{The average countings per segments for each replica}
\usage{
counts.per.seg(list.tau, list.data)
}
\arguments{
\item{list.tau}{list of changepoints for each replica, for a given penalty value.}

\item{list.data}{list of dataset for each replica.}
}
\value{
a matrix containing the average count for each segment (column), for each replica(row).
}
\description{
The average countings per segments for each replica
}
\examples{
l.d1 <- log.transform(dataset1)
seg_rob1      <- Rob_seg.std(x = l.d1,  loss = "Outlier", lambda = 25*log(length(l.d1)), lthreshold=3)
tau1           <- seg_rob1$t.est
l.d2 <- log.transform(dataset2)
seg_rob2      <- Rob_seg.std(x = l.d2,  loss = "Outlier", lambda = 25*log(length(l.d1)), lthreshold=3)
tau2           <- seg_rob2$t.est
l.d3 <- log.transform(dataset3)
seg_rob3      <- Rob_seg.std(x = l.d3,  loss = "Outlier", lambda = 25*log(length(l.d1)), lthreshold=3)
tau3           <- seg_rob3$t.est
l.data <- list(dataset1,dataset2,dataset3)
l.tau <- list(tau1,tau2,tau3)
cps <- counts.per.seg(list.tau=l.tau,list.data=l.data)
}
