% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{CROPS.RFPOP_}
\alias{CROPS.RFPOP_}
\title{Segmenting a dataset using rob_seg, over a penalty range scanned by a CROPS algorithm}
\usage{
CROPS.RFPOP_(data, min_pen = 5, max_pen = 20, lthreshold = 3)
}
\arguments{
\item{data}{The dataset, a vector}

\item{min_pen}{minimum value of the penalty range}

\item{max_pen}{maximum value of the penalty range}

\item{lthreshold}{the threshold used to detect and rescale the outliers among the dataset (3 by default)}
}
\value{
A list: respectively a matrix of penalties (the 2nd line contains the intermediate penalties),
 a list of segmentations for each intermediate lambda and
 a list of smt for each intermediate lambda (they are obviously all the same...)
}
\description{
Segmenting a dataset using rob_seg, over a penalty range scanned by a CROPS algorithm
}
\examples{
log.t1 <- log.transform(dataRNA[,1])
crops.out  <- CROPS.RFPOP_(log.t1,min_pen = 5,max_pen = 10,lthreshold = 3)
View(crops.out)

}
