% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{getIndexLambda}
\alias{getIndexLambda}
\title{returning the index of a penalty among a vector of intermediate penalties}
\usage{
getIndexLambda(lambda, o.crops)
}
\arguments{
\item{lambda}{a penalty value}

\item{o.crops}{the output object from CROPS.RFPOP_}
}
\value{
the index of the penalty, an integer
}
\description{
returning the index of a penalty among a vector of intermediate penalties
}
\examples{
log.t1 <- log.transform(dataRNA[,1])
crops.out  <- CROPS.RFPOP_(log.t1,min_pen = 5,max_pen = 10,lthreshold = 3)
getInd <- getIndex(lambda = 10*log(n), o.crops = crops.out)
getInd

}
