% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{nid.penalties}
\alias{nid.penalties}
\title{Computing NID depending on penalty}
\usage{
nid.penalties(list.seg)
}
\arguments{
\item{list.seg}{an "SEGMENTATIONS" object}
}
\value{
a matrix of results: the lines represent the observations (the replicas), the "avant-dernière" line: the mean of them. The last line: the intermediate penalties. The columns represent the nid value for each penalty.
}
\description{
Computing NID depending on penalty
}
\examples{
l.d1     <- log.transform(dataset1)
l.d2     <- log.transform(dataset2)
l.d3     <- log.transform(dataset3)
l.data   <- list(l.d1,l.d2,l.d3)
list.res <- seg.func(func="rob_seg",list.rna.data=l.data,penalty_range=c(10,100))
nid.res  <- nid.penalties(list.res)

}
