% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{plot.MSE}
\alias{plot.MSE}
\title{Ploting the mse depending on the penalties}
\usage{
\method{plot}{MSE}(mse.res)
}
\arguments{
\item{mse.res}{a "MSE" object}
}
\value{
a graph...
}
\description{
Ploting the mse depending on the penalties
}
\examples{
plot(mse.res)
}
