% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{plot.NID}
\alias{plot.NID}
\title{Ploting the nid depending on the penalties}
\usage{
\method{plot}{NID}(nid.res)
}
\arguments{
\item{mse.res}{a "NID" object}
}
\value{
a graph...
}
\description{
Ploting the nid depending on the penalties
}
\examples{
plot(nid.res)

}
