% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{seg.criteria}
\alias{seg.criteria}
\title{Computing different criterion for RNAs segmentations}
\usage{
seg.criteria(list.seg, criterion)
}
\arguments{
\item{list.seg}{A "SEGMENTATIONS" object}

\item{criterion}{A string indicating the selected criterion: "MSE" or "NID" so far}
}
\value{
the selected criterion and a plot
}
\description{
Computing different criterion for RNAs segmentations
}
\examples{
l.d1     <- log.transform(dataset1)
l.d2     <- log.transform(dataset2)
l.d3     <- log.transform(dataset3)
l.data   <- list(l.d1,l.d2,l.d3)
list.res <- seg.func(func="rob_seg",list.rna.data=l.data,penalty_range=c(10,100))
crit.res <- seg.criteria(list.res, criterion="MSE")
crit.res2 <- seg.criteria(list.res, criterion="NID")

}
