% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_criteria.R
\name{seg.func}
\alias{seg.func}
\title{Segmenting a list of dataset for a given segmentation algo and a penalty range}
\usage{
seg.func(list.rna.data, penalty_range, mc.cores = 1)
}
\arguments{
\item{list.rna.data}{a list of dataset}

\item{penalty_range}{a vector of length 2. Respectively. The min and max penalties}

\item{mc.cores}{a paramater for mclapply, 1 by default... (CF the appropriate documentation about mcapply...)}
}
\value{
a list of values: a list of tau for each beta, a list of smt for each dataset, a list of intermediate penalties.
}
\description{
Segmenting a list of dataset for a given segmentation algo and a penalty range
}
\examples{
l.d1     <- log.transform(dataset1)
l.d2     <- log.transform(dataset2)
l.d3     <- log.transform(dataset3)
l.data   <- list(l.d1,l.d2,l.d3)
list.res <- seg.func(list.rna.data=l.data,penalty_range=c(10,100),mc.cores = 1)
View  (list.res)

}
